+++ src/vivarium/framework/population/manager.py
@@ -1,4 +1,3 @@
-# mypy: ignore-errors
 """
 ======================
 The Population Manager
@@ -8,18 +7,26 @@ The manager and :ref:`builder <builder_concept>` interface for the
 :ref:`population management system <population_concept>`.
 
 """
+from __future__ import annotations
 
+from collections.abc import Callable, Sequence
+from dataclasses import dataclass
 from types import MethodType
-from typing import Any, Callable, Dict, List, NamedTuple, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any
 
 import pandas as pd
 
 from vivarium.framework.population.exceptions import PopulationError
 from vivarium.framework.population.population_view import PopulationView
-from vivarium.manager import Manager
+from vivarium.manager import Interface, Manager
+from vivarium.types import ClockStepSize, ClockTime
+
+if TYPE_CHECKING:
+    from vivarium.framework.engine import Builder
 
 
-class SimulantData(NamedTuple):
+@dataclass
+class SimulantData:
     """Data to help components initialize simulants.
 
     Any time simulants are added to the simulation, each initializer is called
@@ -29,25 +36,27 @@ class SimulantData(NamedTuple):
     """
 
     #: The index representing the new simulants being added to the simulation.
-    index: pd.Index
+    index: pd.Index[int]
     #: A dictionary of extra data passed in by the component creating the
     #: population.
-    user_data: Dict[str, Any]
+    user_data: dict[str, Any]
     #: The time when the simulants enter the simulation.
-    creation_time: pd.Timestamp
+    creation_time: ClockTime
     #: The span of time over which the simulants are created.  Useful for,
     #: e.g., distributing ages over the window.
-    creation_window: pd.Timedelta
+    creation_window: ClockStepSize
 
 
 class InitializerComponentSet:
     """Set of unique components with population initializers."""
 
-    def __init__(self):
-        self._components = {}
-        self._columns_produced = {}
+    def __init__(self) -> None:
+        self._components: dict[str, list[str]] = {}
+        self._columns_produced: dict[str, str] = {}
 
-    def add(self, initializer: Callable, columns_produced: List[str]) -> None:
+    def add(
+        self, initializer: Callable[[SimulantData], None], columns_produced: list[str]
+    ) -> None:
         """Adds an initializer and columns to the set, enforcing uniqueness.
 
         Parameters
@@ -108,10 +117,10 @@ class InitializerComponentSet:
             self._columns_produced[column] = component_name
         self._components[component_name] = columns_produced
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return repr(self._components)
 
-    def __str__(self):
+    def __str__(self) -> str:
         return str(self._components)
 
 
@@ -126,8 +135,15 @@ class PopulationManager(Manager):
         },
     }
 
-    def __init__(self):
-        self._population = None
+    @property
+    def population(self) -> pd.DataFrame:
+        """The current population state table."""
+        if self._population is None:
+            raise PopulationError("Population has not been initialized.")
+        return self._population
+
+    def __init__(self) -> None:
+        self._population: pd.DataFrame | None = None
         self._initializer_components = InitializerComponentSet()
         self.creating_initial_population = False
         self.adding_simulants = False
@@ -138,11 +154,11 @@ class PopulationManager(Manager):
     ############################
 
     @property
-    def name(self):
+    def name(self) -> str:
         """The name of this component."""
         return "population_manager"
 
-    def setup(self, builder):
+    def setup(self, builder: Builder) -> None:
         """Registers the population manager with other vivarium systems."""
         self.clock = builder.time.clock()
         self.step_size = builder.time.step_size()
@@ -165,30 +181,23 @@ class PopulationManager(Manager):
         )
 
         self.register_simulant_initializer(
-            self.on_initialize_simulants, creates_columns=["tracked"]
+            self.on_initialize_simulants, creates_columns="tracked"
         )
-        self._view = self.get_view(["tracked"])
+        self._view = self.get_view("tracked")
 
-    def on_initialize_simulants(self, pop_data: SimulantData):
+    def on_initialize_simulants(self, pop_data: SimulantData) -> None:
         """Adds a ``tracked`` column to the state table for new simulants."""
         status = pd.Series(True, index=pop_data.index)
         self._view.update(status)
 
-    @property
-    def columns(self) -> List[str]:
-        """The columns that currently exist in the state table."""
-        return list(self._population.columns)
-
-    def __repr__(self):
+    def __repr__(self) -> str:
         return "PopulationManager()"
 
     ###########################
     # Builder API and helpers #
     ###########################
 
-    def get_view(
-        self, columns: Union[List[str], Tuple[str]], query: Optional[str] = None
-    ) -> PopulationView:
+    def get_view(self, columns: str | Sequence[str], query: str = "") -> PopulationView:
         """Get a time-varying view of the population state table.
 
         The requested population view can be used to view the current state or
@@ -235,11 +244,12 @@ class PopulationManager(Manager):
         )
         return view
 
-    def _get_view(
-        self, columns: Union[List[str], Tuple[str]], query: Optional[str] = None
-    ) -> PopulationView:
+    def _get_view(self, columns: str | Sequence[str], query: str) -> PopulationView:
+        if isinstance(columns, str):
+            columns = [columns]
+
         if columns and "tracked" not in columns:
-            if query is None:
+            if not query:
                 query = "tracked == True"
             elif "tracked" not in query:
                 query += " and tracked == True"
@@ -248,11 +258,11 @@ class PopulationManager(Manager):
 
     def register_simulant_initializer(
         self,
-        initializer: Callable,
-        creates_columns: List[str] = (),
-        requires_columns: List[str] = (),
-        requires_values: List[str] = (),
-        requires_streams: List[str] = (),
+        initializer: Callable[[SimulantData], None],
+        creates_columns: str | Sequence[str] = (),
+        requires_columns: str | Sequence[str] = (),
+        requires_values: str | Sequence[str] = (),
+        requires_streams: str | Sequence[str] = (),
     ) -> None:
         """Marks a source of initial state information for new simulants.
 
@@ -262,26 +272,35 @@ class PopulationManager(Manager):
             A callable that adds or updates initial state information about
             new simulants.
         creates_columns
-            A list of the state table columns that the given initializer
-            provides the initial state information for.
+            The state table columns that the given initializer provides the
+            initial state information for.
         requires_columns
-            A list of the state table columns that already need to be present
-            and populated in the state table before the provided initializer
-            is called.
+            The state table columns that already need to be present and
+            populated in the state table before the provided initializer is
+            called.
         requires_values
-            A list of the value pipelines that need to be properly sourced
-            before the provided initializer is called.
+            The value pipelines that need to be properly sourced before the
+            provided initializer is called.
         requires_streams
-            A list of the randomness streams necessary to initialize the
-            simulant attributes.
+            The randomness streams necessary to initialize the simulant
+            attributes.
         """
-        self._initializer_components.add(initializer, creates_columns)
+        if isinstance(creates_columns, str):
+            creates_columns = [creates_columns]
+        if isinstance(requires_columns, str):
+            requires_columns = [requires_columns]
+        if isinstance(requires_values, str):
+            requires_values = [requires_values]
+        if isinstance(requires_streams, str):
+            requires_streams = [requires_streams]
+
+        self._initializer_components.add(initializer, list(creates_columns))
         dependencies = (
             [f"column.{name}" for name in requires_columns]
             + [f"value.{name}" for name in requires_values]
             + [f"stream.{name}" for name in requires_streams]
         )
-        if creates_columns != ["tracked"]:
+        if "tracked" not in creates_columns:
             # The population view itself uses the tracked column, so include
             # to be safe.
             dependencies += ["column.tracked"]
@@ -289,7 +308,7 @@ class PopulationManager(Manager):
             "column", list(creates_columns), initializer, dependencies
         )
 
-    def get_simulant_creator(self) -> Callable[[int, Optional[Dict[str, Any]]], pd.Index]:
+    def get_simulant_creator(self) -> Callable[[int, dict[str, Any] | None], pd.Index[int]]:
         """Gets a function that can generate new simulants.
 
         The creator function takes the number of simulants to be created as it's
@@ -308,8 +327,8 @@ class PopulationManager(Manager):
         return self._create_simulants
 
     def _create_simulants(
-        self, count: int, population_configuration: Dict[str, Any] = None
-    ) -> pd.Index:
+        self, count: int, population_configuration: dict[str, Any] | None = None
+    ) -> pd.Index[int]:
         population_configuration = (
             population_configuration if population_configuration else {}
         )
@@ -353,7 +372,7 @@ class PopulationManager(Manager):
         return pop
 
 
-class PopulationInterface:
+class PopulationInterface(Interface):
     """Provides access to the system for reading and updating the population.
 
     The most important aspect of the simulation state is the ``population
@@ -379,9 +398,7 @@ class PopulationInterface:
     def __init__(self, manager: PopulationManager):
         self._manager = manager
 
-    def get_view(
-        self, columns: Union[List[str], Tuple[str]], query: Optional[str] = None
-    ) -> PopulationView:
+    def get_view(self, columns: str | Sequence[str], query: str = "") -> PopulationView:
         """Get a time-varying view of the population state table.
 
         The requested population view can be used to view the current state or
@@ -413,7 +430,7 @@ class PopulationInterface:
         """
         return self._manager.get_view(columns, query)
 
-    def get_simulant_creator(self) -> Callable[[int, Optional[Dict[str, Any]]], pd.Index]:
+    def get_simulant_creator(self) -> Callable[[int, dict[str, Any] | None], pd.Index[int]]:
         """Gets a function that can generate new simulants.
 
         The creator function takes the number of simulants to be created as it's
@@ -434,10 +451,10 @@ class PopulationInterface:
     def initializes_simulants(
         self,
         initializer: Callable[[SimulantData], None],
-        creates_columns: List[str] = (),
-        requires_columns: List[str] = (),
-        requires_values: List[str] = (),
-        requires_streams: List[str] = (),
+        creates_columns: str | Sequence[str] = (),
+        requires_columns: str | Sequence[str] = (),
+        requires_values: str | Sequence[str] = (),
+        requires_streams: str | Sequence[str] = (),
     ) -> None:
         """Marks a source of initial state information for new simulants.
 
@@ -447,17 +464,17 @@ class PopulationInterface:
             A callable that adds or updates initial state information about
             new simulants.
         creates_columns
-            A list of the state table columns that the given initializer
+            The state table columns that the given initializer
             provides the initial state information for.
         requires_columns
-            A list of the state table columns that already need to be present
+            The state table columns that already need to be present
             and populated in the state table before the provided initializer
             is called.
         requires_values
-            A list of the value pipelines that need to be properly sourced
+            The value pipelines that need to be properly sourced
             before the provided initializer is called.
         requires_streams
-            A list of the randomness streams necessary to initialize the
+            The randomness streams necessary to initialize the
             simulant attributes.
         """
         self._manager.register_simulant_initializer(