+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+**3.0.7 - 09/24/24**
+
+  - Fix mypy errors: vivarium/framework/lifecycle.py
+  - Fix mypy errors: vivarium/framework/population/manager.py
+  - Fix mypy errors: vivarium/framework/population/population_view.py
+  - Enable population manager and population view methods to accept strings
+  - Fix mypy errors: vivarium/framework/plugins.py
+
 **3.0.6 - 09/20/24**
 
   - Fix mypy errors: vivarium/framework/randomness/index_map.py